админ-панель -> Головна

+++ admin/controller/Gallery/del.php
@@ -3,7 +3,9 @@
 session_start();
 $connect = mysqli_connect("localhost", "root", "", "rio_barbershop");
 
-
+if(!$connect) {
+    die("Mysql no work,pleas check connect mysql");
+}
 $id = $_POST['id'];
 
     mysqli_query($connect, " DELETE FROM gallery WHERE id = '$id'");
